+++ debian/debian/rules
@@ -10,6 +10,7 @@ override_dh_auto_install:
 
 override_dh_installsystemd:
 	dh_installsystemd --no-start -r --no-restart-after-upgrade --name=ix-boot-core
+	dh_installsystemd --no-start -r --no-restart-after-upgrade --name=ix-cgroups
 	dh_installsystemd --no-start -r --no-restart-after-upgrade --name=ix-conf
 	dh_installsystemd --no-start -r --no-restart-after-upgrade --name=ix-etc
 	dh_installsystemd --no-start -r --no-restart-after-upgrade --name=ix-freebsd-to-scale-update
