+++ src/middlewared/middlewared/plugins/kubernetes_linux/lifecycle.py
@@ -12,10 +12,7 @@ from middlewared.service import CallError, private, Service
 from middlewared.utils import run
 
 from .k8s.config import reinitialize_config
-from .utils import (
-    CGROUP_ROOT_PATH, get_available_controllers_for_consumption, RE_CGROUP_CONTROLLERS,
-    update_available_controllers_for_consumption,
-)
+
 
 START_LOCK = asyncio.Lock()
 
@@ -198,35 +195,6 @@ class KubernetesService(Service):
         }
 
     @private
-    def ensure_cgroups_are_setup(self):
-        # Logic copied over from kubernetes
-        # https://github.com/kubernetes/kubernetes/blob/08fbe92fa76d35048b4b4891b41fc6912e689cc7/
-        # pkg/kubelet/cm/cgroup_manager_linux.go#L238
-        supported_controllers = {'cpu', 'cpuset', 'memory', 'hugetlb', 'pids'}
-        system_supported_controllers_path = os.path.join(CGROUP_ROOT_PATH, 'cgroup.controllers')
-        try:
-            with open(system_supported_controllers_path, 'r') as f:
-                available_controllers = set(RE_CGROUP_CONTROLLERS.findall(f.read()))
-        except FileNotFoundError:
-            raise CallError(
-                'Unable to determine available cgroup controllers as '
-                f'{system_supported_controllers_path!r} does not exist'
-            )
-
-        needed_controllers = supported_controllers & available_controllers
-        available_controllers_for_consumption = get_available_controllers_for_consumption()
-        if missing_controllers := needed_controllers - available_controllers_for_consumption:
-            # If we have missing controllers, lets try adding them to subtree control
-            available_controllers_for_consumption = update_available_controllers_for_consumption(missing_controllers)
-
-        missing_controllers = needed_controllers - available_controllers_for_consumption
-        if missing_controllers:
-            raise CallError(
-                f'Missing {", ".join(missing_controllers)!r} cgroup controller(s) '
-                'which are required for apps to function'
-            )
-
-    @private
     async def validate_k8s_fs_setup(self):
         config = await self.middleware.call('kubernetes.config')
         if not await self.middleware.call('pool.query', [['name', '=', config['pool']]]):
@@ -272,7 +240,6 @@ class KubernetesService(Service):
         if errors:
             raise CallError(str(errors))
 
-        await self.middleware.call('kubernetes.ensure_cgroups_are_setup')
         await self.middleware.call('k8s.migration.scale_version_check')
 
     @private
